QuoteFinder/requirements.txt
openai-whisper>=20231117
torch>=2.0.0
ffmpeg-python>=0.2.0
rapidfuzz>=3.0.0
